refactor/all: add crypto module
The module will contain all cryptographic primitives used by routing. Currently it contains only `hash` function (alias for tiny_keccak::sha3_256) and `Digest256` type.
