+++ src/routing_message_filter.rs
@@ -6,12 +6,9 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
-use crate::id::PublicId;
-use crate::message_filter::MessageFilter;
-use crate::messages::RoutingMessage;
+use crate::{crypto, id::PublicId, message_filter::MessageFilter, messages::RoutingMessage};
 use lru_time_cache::LruCache;
 use maidsafe_utilities::serialisation::serialise;
-use safe_crypto;
 use serde::Serialize;
 use std::fmt::Debug;
 use std::time::Duration;
@@ -80,7 +77,7 @@ impl Default for RoutingMessageFilter {
 
 fn hash<T: Serialize + Debug>(msg: &T) -> Option<Digest> {
     if let Ok(msg_bytes) = serialise(msg) {
-        Some(safe_crypto::hash(&msg_bytes))
+        Some(crypto::hash(&msg_bytes))
     } else {
         trace!("Tried to filter oversized routing message: {:?}", msg);
         None
