+++ src/states/elder/mod.rs
@@ -19,6 +19,7 @@ use crate::{
         SectionInfo, SectionInfoSigPayload, SectionKeyInfo, SendAckMessagePayload,
     },
     config_handler,
+    crypto::Digest256,
     error::{BootstrapResponseError, InterfaceError, RoutingError},
     event::Event,
     id::{FullId, PublicId},
@@ -31,7 +32,6 @@ use crate::{
     routing_message_filter::{FilteringResult, RoutingMessageFilter},
     routing_table::Error as RoutingTableError,
     routing_table::{Authority, Prefix, Xorable, DEFAULT_PREFIX},
-    sha3::Digest256,
     signature_accumulator::SignatureAccumulator,
     state_machine::Transition,
     time::{Duration, Instant},
